Vault Project/requirements.txt
python-dotenv
PyJWT